generation/requirements.txt
certifi==2019.11.28
chardet==3.0.4
cwlgen==0.4.0
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
nodeenv==1.3.5
requests==2.23.0
ruamel.yaml==0.16.5
six==1.14.0
smmap==3.0.1
urllib3==1.25.8
